chore: migrate types to fetch and updates foo.com to example.com based on review feedback
